Better const name and some int was replaced by double

+++ headers/cub3d.function.h
@@ -6,7 +6,7 @@
 /*   By: emaillet <emaillet@student.42lehavre.fr    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/19 09:31:49 by emaillet          #+#    #+#             */
-/*   Updated: 2025/05/21 13:11:03 by emaillet         ###   ########.fr       */
+/*   Updated: 2025/05/21 14:10:03 by emaillet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,7 +60,7 @@ void		find_max(char **map, int *x, int *y);
 void		map_size(t_c3_data *data);
 
 //Utils
-t_trigo		trigo(t_c3_data *data, int x, int y);
+t_trigo		trigo(t_c3_data *data, double x, double y);
 int			ft_tablen(char **tab);
 void		print_tab(char **tab);
 char		**lst_to_tab(t_list *lst);
